ticket/requirements.txt
asgiref==3.4.1
astroid==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
Django==4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.8.2
httpie==2.6.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
platformdirs==2.4.0
psycopg2==2.9.2
Pygments==2.10.0
pylint==2.12.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
